requirements.txt
lxml==4.9.3
python-docx==0.8.11
colorama==0.4.6
exceptiongroup==1.1.3
iniconfig==2.0.0
lxml==4.9.3
packaging==23.2
pluggy==1.3.0
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pytest==7.4.2
python-docx==0.8.11
tomli==2.0.1
